# docs/swagger.json
{
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a simple currency API",
        "title": "Currency API",
        "contact": {
            "name": "Telegram",
            "url": "https://t.me/ilyas_amantaev"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/currency/save/{date}": {
            "get": {
                "description": "Fetch and store rates by date",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rates"
                ],
                "summary": "Fetch and store rates",
                "parameters": [
                    {
                        "type": "string",
                        "description": "date. Example: '01-01-2022'",
                        "name": "date",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_kartochnik010_outstaff-task_internal_pkg_js.JSON"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_kartochnik010_outstaff-task_internal_pkg_js.JSON"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/currency/{date}/{code}": {
            "get": {
                "description": "Get rates by date",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rates"
                ],
                "summary": "Get rates",
                "parameters": [
                    {
                        "type": "string",
                        "description": "date. Example: '01-01-2022'",
                        "name": "date",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "code. Example: 'USD'",
                        "name": "code",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_kartochnik010_outstaff-task_internal_pkg_js.JSON"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "rates": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_kartochnik010_outstaff-task_internal_domain_models.Rate"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_kartochnik010_outstaff-task_internal_pkg_js.JSON"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_kartochnik010_outstaff-task_internal_domain_models.Rate": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "github_com_kartochnik010_outstaff-task_internal_pkg_js.JSON": {
            "type": "object",
            "additionalProperties": true
        }
    }
}